requirements.txt
psycopg2==2.5.2

Django==1.6.1
south==0.8.4
django-crispy-forms==1.3.2
linaro-django-pagination==2.0.2
